20211220(월) 글쓰기페이지 --> DB insert test 오류 : list page에 data 안나옴 ( NullPointerException )

+++ BookSales/src/main/java/com/board/domain/BoardDTO.java
@@ -11,7 +11,7 @@ public class BoardDTO {
 	
 		//db칼럼명은 대문자 java 소문자로 
 	
-	private int post_number;    
+	private Long post_number;    
 	//POST_NUMBER 게시번호 
 	
 	private int board_number;
